refactor：将分页条件查询提取为一个公共方法

+++ src/main/java/com/fire/phenix/devops/mapper/SysMenuMapper.java
@@ -19,5 +19,5 @@ public interface SysMenuMapper extends BaseMapper<SysMenu> {
      * @param menu 菜单项信息
      * @return 修改的条目数
      */
-    Long updatePartFieldById(@Param("id") Long id, @Param("menu") SysMenu menu);
+    int updatePartFieldById(@Param("id") Long id, @Param("menu") SysMenu menu);
 }
